# server/swaggerDist/pacomeswagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a Swagger documentation of endpoints for Pacome's website backend API.",
    "version": "1.0.0",
    "title": "Pacome's Back-End",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "smbonimpa2011@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:5000",
  "basePath": "/api",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/users/promote/{id}": {
      "patch": {
        "tags": [
          "users"
        ],
        "summary": "promote/demote a user to/from admin",
        "description": "",
        "operationId": "promote",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "x-auth-token",
            "type": "string",
            "required": true
          },
          {
            "name": "id",
            "in": "path",
            "description": "ID of user to promote/demote",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "OK: successfully updated user's credentials.",
            "schema": {
              "$ref": "#/definitions/UserRes"
            }
          },
          "400": {
            "description": "invalid authentication token.",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "401": {
            "description": "no authorized/unauthorized token.",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Error": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string"
        }
      },
      "xml": {
        "name": "Error"
      }
    },
...
    "UserRes": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "membership": {
          "type": "string"
        }
      },
      "xml": {
        "name": "UserRes"
      }
    },
...
  },
...
}